Change the name attributes of the interpreter implementations to class attributes.
This is motivated by a desire to gather up a list of interpreters,
but in general attributes that are constant across all instances of
a class should be class attributes.  There are other examples in these
classes that could be changed, but I won't bother right now since
in practice there's hardly ever more than one instance of these classes
at a time.

It might also make sense to make the base class an ABC, but again not
worth bothering right now.