Provide ctypes.util.find_library replacement that works on Cygwin
ctypes.util.find_library does not work on all platforms (notably cygwin) so
code that uses it (of which there is currently only one place) should import
it from sage.misc.misc rather than ctypes; here we can provide additional
platform support as needed.

Not sure if this is the best place for this; would be good to have a module
specifically for platform compatibility utils.