Fixes to testing sage_setup
Turns out adding all of sage_setup to the 'in_lib' list broke some tests unexpectedly.  In particular the find and clean modules, as well as autogen.pari.doc, which relies on the unicode_literals __future__, which is not loaded if the file is not tested as a lone module.

Also cleaned up the implementation of in_lib to be a bit robust.  Every package listed is treated as 'in_lib' as well as submodules of those packages.