A fairly significant restructuring:  Much of the preparation that was performed in the metaclass is now moved into the __init__ of the Makefile base class.
Although it was nice to have sanity checking happen at the class
definition time itself (and maybe we can add more back in later in
certain cases) there are too many problems with this.  Most notably,
dependence on runtime factors like wildcard expansions--that should
be happening when we actually run the makefile, not just when we
define it, when the environment might be totally wrong.

An upshot to this is that it also significantly simplifies the
metaclass which is also good for comprehension.

This added some other small enhancements, such as that all
rules get a name now (even unbound ones).